feat: refactor of blockchain service out of content-watcher
