+++ apps/content-watcher/src/api.module.ts
@@ -3,7 +3,7 @@ import { EventEmitterModule } from '@nestjs/event-emitter';
 import { ScheduleModule } from '@nestjs/schedule';
 import { ApiService } from './api.service';
 import { HealthController, ScanControllerV1, SearchControllerV1, WebhookControllerV1 } from './controllers';
-import { BlockchainModule } from '#content-watcher-lib/blockchain/blockchain.module';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 import { CrawlerModule } from '#content-watcher-lib/crawler/crawler.module';
 import { IPFSProcessorModule } from '#content-watcher-lib/ipfs/ipfs.module';
 import { PubSubModule } from '#content-watcher-lib/pubsub/pubsub.module';
@@ -13,7 +13,7 @@ import { CacheModule } from '#cache/cache.module';
 import cacheConfig, { ICacheConfig } from '#cache/cache.config';
 import { ConfigModule } from '@nestjs/config';
 import apiConfig from './api.config';
-import blockchainConfig from '#content-watcher-lib/blockchain/blockchain.config';
+import blockchainConfig from '#blockchain/blockchain.config';
 import queueConfig, { QueueModule } from '#queue';
 import ipfsConfig from '#content-watcher-lib/ipfs/ipfs.config';
 import scannerConfig from '#content-watcher-lib/scanner/scanner.config';
