--- libs/content-watcher-lib/src/blockchain/blockchain.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BlockchainService } from './blockchain.service';
-import { ChainEventProcessorService } from './chain-event-processor.service';
-
-@Module({
-  imports: [],
-  controllers: [],
-  providers: [BlockchainService, ChainEventProcessorService],
-  exports: [BlockchainService, ChainEventProcessorService],
-})
-export class BlockchainModule {}
