--- libs/content-watcher-lib/src/blockchain/blockchain.service.ts
@@ -1,130 +1,0 @@
-/* eslint-disable no-underscore-dangle */
-import { Inject, Injectable, Logger, OnApplicationBootstrap, OnApplicationShutdown } from '@nestjs/common';
-import { ApiPromise, ApiRx, HttpProvider, WsProvider } from '@polkadot/api';
-import { firstValueFrom } from 'rxjs';
-import { options } from '@frequency-chain/api-augment';
-import { KeyringPair } from '@polkadot/keyring/types';
-import { BlockHash, BlockNumber, SignedBlock } from '@polkadot/types/interfaces';
-import { SubmittableExtrinsic } from '@polkadot/api/types';
-import { AnyNumber, ISubmittableResult } from '@polkadot/types/types';
-import { Extrinsic } from './extrinsic';
-import blockchainConfig, { IBlockchainConfig } from './blockchain.config';
-
-@Injectable()
-export class BlockchainService implements OnApplicationBootstrap, OnApplicationShutdown {
-  public api: ApiRx;
-
-  public apiPromise: ApiPromise;
-
-  private config: IBlockchainConfig;
-
-  private logger: Logger;
-
-  public async onApplicationBootstrap() {
-    const providerUrl = this.config.frequencyUrl!;
-    let provider: any;
-    if (/^ws/.test(providerUrl.toString())) {
-      provider = new WsProvider(providerUrl.toString());
-    } else if (/^http/.test(providerUrl.toString())) {
-      provider = new HttpProvider(providerUrl.toString());
-    } else {
-      this.logger.error(`Unrecognized chain URL type: ${providerUrl.toString()}`);
-      throw new Error('Unrecognized chain URL type');
-    }
-    this.api = await firstValueFrom(ApiRx.create({ provider, ...options }));
-    this.apiPromise = await ApiPromise.create({ provider, ...options });
-    await Promise.all([firstValueFrom(this.api.isReady), this.apiPromise.isReady]);
-    this.logger.log('Blockchain API ready.');
-  }
-
-  public async isReady(): Promise<boolean> {
-    await this.apiPromise.isReady;
-    return true;
-  }
-
-  public async onApplicationShutdown(_signal?: string | undefined) {
-    const promises: Promise<any>[] = [];
-    if (this.api) {
-      promises.push(this.api.disconnect());
-    }
-
-    if (this.apiPromise) {
-      promises.push(this.apiPromise.disconnect());
-    }
-    await Promise.all(promises);
-  }
-
-  constructor(@Inject(blockchainConfig.KEY) config: IBlockchainConfig) {
-    this.config = config;
-    this.logger = new Logger(this.constructor.name);
-  }
-
-  public getBlockHash(block: BlockNumber | AnyNumber): Promise<BlockHash> {
-    return firstValueFrom(this.api.rpc.chain.getBlockHash(block));
-  }
-
-  public getBlock(block?: BlockHash): Promise<SignedBlock> {
-    return firstValueFrom(this.api.rpc.chain.getBlock(block));
-  }
-
-  public async getLatestFinalizedBlockHash(): Promise<BlockHash> {
-    return (await this.apiPromise.rpc.chain.getFinalizedHead()) as BlockHash;
-  }
-
-  public async getLatestFinalizedBlockNumber(): Promise<number> {
-    return (await this.apiPromise.rpc.chain.getBlock()).block.header.number.toNumber();
-  }
-
-  public async getBlockNumberForHash(hash: string): Promise<number | undefined> {
-    const block = await this.apiPromise.rpc.chain.getBlock(hash);
-    if (block) {
-      return block.block.header.number.toNumber();
-    }
-
-    this.logger.error(`No block found corresponding to hash ${hash}`);
-    return undefined;
-  }
-
-  public createType(type: string, ...args: (any | undefined)[]) {
-    return this.api.registry.createType(type, ...args);
-  }
-
-  public createExtrinsicCall(
-    { pallet, extrinsic }: { pallet: string; extrinsic: string },
-    ...args: (any | undefined)[]
-  ): SubmittableExtrinsic<'rxjs', ISubmittableResult> {
-    return this.api.tx[pallet][extrinsic](...args);
-  }
-
-  public createExtrinsic(
-    { pallet, extrinsic }: { pallet: string; extrinsic: string },
-    { eventPallet, event }: { eventPallet?: string; event?: string },
-    keys: KeyringPair,
-    ...args: (any | undefined)[]
-  ): Extrinsic {
-    const targetEvent = eventPallet && event ? this.api.events[eventPallet][event] : undefined;
-    return new Extrinsic(this.api, this.api.tx[pallet][extrinsic](...args), keys, targetEvent);
-  }
-
-  public rpc(pallet: string, rpc: string, ...args: (any | undefined)[]): Promise<any> {
-    return this.apiPromise.rpc[pallet][rpc](...args);
-  }
-
-  public query(pallet: string, extrinsic: string, ...args: (any | undefined)[]): Promise<any> {
-    return args ? this.apiPromise.query[pallet][extrinsic](...args) : this.apiPromise.query[pallet][extrinsic]();
-  }
-
-  public async queryAt(
-    blockHash: BlockHash,
-    pallet: string,
-    extrinsic: string,
-    ...args: (any | undefined)[]
-  ): Promise<any> {
-    const newApi = await this.apiPromise.at(blockHash);
-    return newApi.query[pallet][extrinsic](...args);
-  }
-
-  public async getNonce(account: Uint8Array): Promise<number> {
-    return this.rpc('system', 'accountNextIndex', account);
-  }
-}
