+++ libs/content-watcher-lib/src/crawler/crawler.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ScheduleModule } from '@nestjs/schedule';
 import { CrawlerService } from './crawler.service';
-import { BlockchainModule } from '../blockchain/blockchain.module';
+import { BlockchainModule } from '#blockchain/blockchain.module';
+import { ChainEventProcessorModule } from '#content-watcher-lib/utils/chain-event-process.module';
 
 @Module({
-  imports: [BlockchainModule, ScheduleModule],
+  imports: [BlockchainModule, ScheduleModule, ChainEventProcessorModule],
   controllers: [],
   providers: [CrawlerService],
   exports: [CrawlerService],
