+++ libs/content-watcher-lib/src/crawler/crawler.service.ts
@@ -8,8 +8,8 @@ import { ContentWatcherQueues as QueueConstants } from '#types/constants/queue.c
 import { ChainWatchOptionsDto } from '#types/dtos/content-watcher/chain.watch.dto';
 import { BaseConsumer } from '../utils/base-consumer';
 import { ContentSearchRequestDto } from '#types/dtos/content-watcher/content-search-request.dto';
-import { ChainEventProcessorService } from '../blockchain/chain-event-processor.service';
-import { BlockchainService } from '../blockchain/blockchain.service';
+import { ChainEventProcessorService } from '../utils/chain-event-processor.service';
+import { BlockchainService } from '#blockchain/blockchain.service';
 
 const CRAWLER_BLOCK_CHUNK_SIZE = 500;
