+++ libs/content-watcher-lib/src/ipfs/ipfs.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { BlockchainModule } from '../blockchain/blockchain.module';
+import { BlockchainModule } from '#blockchain/blockchain.module';
 import { IPFSContentProcessor } from './ipfs.dsnp';
 import { IpfsService } from '../utils/ipfs.client';
