+++ libs/content-watcher-lib/src/scanner/scanner.module.ts
@@ -2,10 +2,11 @@ import '@frequency-chain/api-augment';
 import { Module } from '@nestjs/common';
 import { ScheduleModule } from '@nestjs/schedule';
 import { ScannerService } from './scanner.service';
-import { BlockchainModule } from '../blockchain/blockchain.module';
+import { BlockchainModule } from '#blockchain/blockchain.module';
+import { ChainEventProcessorModule } from '#content-watcher-lib/utils/chain-event-process.module';
 
 @Module({
-  imports: [BlockchainModule, ScheduleModule],
+  imports: [BlockchainModule, ScheduleModule, ChainEventProcessorModule],
   controllers: [],
   providers: [ScannerService],
   exports: [ScannerService],
