+++ libs/content-watcher-lib/src/utils/chain-event-process.module.ts
@@ -1,0 +1,10 @@
+import { BlockchainModule } from '#blockchain/blockchain.module';
+import { Module } from '@nestjs/common';
+import { ChainEventProcessorService } from './chain-event-processor.service';
+
+@Module({
+  imports: [BlockchainModule],
+  providers: [ChainEventProcessorService],
+  exports: [ChainEventProcessorService],
+})
+export class ChainEventProcessorModule {}
