+++ libs/content-watcher-lib/src/utils/chain-event-processor.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { BlockchainService } from './blockchain.service';
+import { BlockchainService } from '#blockchain/blockchain.service';
 import { ChainWatchOptionsDto } from '#types/dtos/content-watcher/chain.watch.dto';
 import { ApiDecoration } from '@polkadot/api/types';
 import { FrameSystemEventRecord } from '@polkadot/types/lookup';
@@ -22,7 +22,7 @@ export class ChainEventProcessorService {
     if (blockHash.isEmpty) {
       return [];
     }
-    const apiAt = await this.blockchainService.apiPromise.at(blockHash);
+    const apiAt = await this.blockchainService.api.at(blockHash);
     const events = await apiAt.query.system.events();
     return this.getMessagesFromEvents(apiAt, blockNumber, events, filter);
   }
@@ -54,7 +54,7 @@ export class ChainEventProcessorService {
         };
 
         let messageResponse: BlockPaginationResponseMessage =
-          await this.blockchainService.apiPromise.rpc.messages.getBySchemaId(schemaId, paginationRequest);
+          await this.blockchainService.api.rpc.messages.getBySchemaId(schemaId, paginationRequest);
         const messages: Vec<MessageResponse> = messageResponse.content;
         while (messageResponse.has_next.toHuman()) {
           paginationRequest = {
@@ -64,10 +64,7 @@ export class ChainEventProcessorService {
             to_block: blockNumber + 1,
           };
           // eslint-disable-next-line no-await-in-loop
-          messageResponse = await this.blockchainService.apiPromise.rpc.messages.getBySchemaId(
-            schemaId,
-            paginationRequest,
-          );
+          messageResponse = await this.blockchainService.api.rpc.messages.getBySchemaId(schemaId, paginationRequest);
           if (messageResponse.content.length > 0) {
             messages.push(...messageResponse.content);
           }
