+++ package.json
@@ -89,7 +89,7 @@
     "test:e2e:graph": "dotenvx run -f env-files/graph.template.env -- jest --testRegex \".e2e-spec.ts\" --detectOpenHandles 'graph*'",
     "test:k6:account": "for t in apps/account-api/k6-test/*.k6.js ; do k6 run ${t} ; done",
     "format": "prettier --write .",
-    "lint": "eslint apps/**/*.[tj]s libs/**/*.[tj]s",
+    "lint": "eslint 'apps/**/*.[tj]s' 'libs/**/*.[tj]s'",
     "lint:account": "eslint 'apps/account*/**/*.[tj]s' 'libs/account-lib/**/*.[tj]s'",
     "lint:content-publishing": "eslint 'apps/content-publishing*/**/*.[tj]s' 'libs/content-publishing-lib/**/*.[tj]s'",
     "lint:content-watcher": "eslint 'apps/content-watcher/**/*.[tj]s' 'libs/content-watcher-lib/**/*.[tj]s'",
